Update game_file.py
Adds a validation function validate_pos_num that takes in three arguments. The prompt for the user, the min value, and max value. This function enables all number inputs to be validated for a specific range and ensure a number in inputted. It is applied in the set up section for the length and width of the board and number of bombs. It is then further used in turn function of the game where it validates the choice of the player for each turn and the row and column picked.